requirements.txt
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1